Rename and adjust files for current location
