+++ keyboards/ymdk/ymd75/rev4/ansi/ansi.c
@@ -16,8 +16,8 @@
 
 // 68keys.c
 
-#include "ymd84a.h"
-#ifdef RGB_MATRIX_ENABLE 
+#include "ansi.h"
+#ifdef RGB_MATRIX_ENABLE
 #define XX NO_LED
 led_config_t g_led_config = {
 { \
@@ -66,35 +66,35 @@ led_config_t g_led_config = {
   /*  {
         {0,   1,   2,   3,   4,   5,   6,   7,    8,  9,   10,  11,  12,  13,  14,  15},
         {16,  17,  18,  19,  20,  21,  22,  23,  24,  25,  26,  27,  28,  29,  30},
-        {31,  32,  33,  34,  35,  36,  37,  38,  39,  40,  41,  42,  43,  44,  45}, 
+        {31,  32,  33,  34,  35,  36,  37,  38,  39,  40,  41,  42,  43,  44,  45},
         {46,  47,  48,  49,  50,  51,  52,  53,  54,  55,  56,  57,  58,  59},
         {60,  61,  62,  63,  64,  65,  66,  67,  68,  69,  70,  71,  72,  73, 74},
         { 75,  76,  77,  78,  79,  80,  81,  82,  83, 84},
     }, {
-        {7, 3}, {21, 3}, {35, 3}, {49, 3}, {63, 3}, {77, 3}, {91, 3}, {105, 3}, {119, 3}, {133, 3}, {147, 3}, {161, 3}, {175, 3}, {189, 3}, {203, 3}, {217, 3}, 
-        {7, 13}, {21, 13}, {35, 13}, {49, 13}, {63, 13}, {77, 13}, {91, 13}, {105, 13}, {119, 13}, {133, 13}, {147, 13}, {161, 13}, {175, 13}, {196, 13}, {217, 13}, 
+        {7, 3}, {21, 3}, {35, 3}, {49, 3}, {63, 3}, {77, 3}, {91, 3}, {105, 3}, {119, 3}, {133, 3}, {147, 3}, {161, 3}, {175, 3}, {189, 3}, {203, 3}, {217, 3},
+        {7, 13}, {21, 13}, {35, 13}, {49, 13}, {63, 13}, {77, 13}, {91, 13}, {105, 13}, {119, 13}, {133, 13}, {147, 13}, {161, 13}, {175, 13}, {196, 13}, {217, 13},
         {10, 24}, {28, 24}, {42, 24}, {56, 24}, {70, 24}, {84, 24}, {98, 24}, {112, 24}, {126, 24}, {140, 24}, {154, 24}, {168, 24}, {182, 24}, {200, 24}, {217, 24},
-        {12, 35}, {32, 35}, {46, 35}, {60, 35}, {74, 35}, {88, 35}, {102, 35}, {116, 35}, {130, 35}, {144, 35}, {158, 35}, {172, 35}, {194, 35}, {217, 35}, 
-        {16, 45}, {38, 45}, {52, 45}, {66, 45}, {80, 45}, {94, 45}, {108, 45}, {122, 45}, {136, 45}, {150, 45}, {164, 45}, {184, 45}, {203, 45}, {217, 45}, 
-        {9, 56}, {26, 56}, {44, 56}, {96, 56}, {147, 56}, {161, 56}, {175, 56}, {189, 56}, {203, 56}, {217, 56}, 
+        {12, 35}, {32, 35}, {46, 35}, {60, 35}, {74, 35}, {88, 35}, {102, 35}, {116, 35}, {130, 35}, {144, 35}, {158, 35}, {172, 35}, {194, 35}, {217, 35},
+        {16, 45}, {38, 45}, {52, 45}, {66, 45}, {80, 45}, {94, 45}, {108, 45}, {122, 45}, {136, 45}, {150, 45}, {164, 45}, {184, 45}, {203, 45}, {217, 45},
+        {9, 56}, {26, 56}, {44, 56}, {96, 56}, {147, 56}, {161, 56}, {175, 56}, {189, 56}, {203, 56}, {217, 56},
         {30, 5}, {60, 5}, {90, 5}, {120, 5}, {150, 5}, {180, 5}, {210, 5},{210, 32},{210, 60},{180, 60},{150, 60},{120, 60}, {90, 60},{60, 60},{30, 60},{30, 32},
     },*/
     {
-        {7, 3}, {21, 3}, {35, 3}, {49, 3}, {63, 3}, {77, 3}, {91, 3}, {105, 3}, {119, 3}, {133, 3}, {147, 3}, {161, 3}, {175, 3}, {189, 3}, {203, 3}, {217, 3}, 
-        {7, 13}, {21, 13}, {35, 13}, {49, 13}, {63, 13}, {77, 13}, {91, 13}, {105, 13}, {119, 13}, {133, 13}, {147, 13}, {161, 13}, {175, 13}, {196, 13}, {217, 13}, 
+        {7, 3}, {21, 3}, {35, 3}, {49, 3}, {63, 3}, {77, 3}, {91, 3}, {105, 3}, {119, 3}, {133, 3}, {147, 3}, {161, 3}, {175, 3}, {189, 3}, {203, 3}, {217, 3},
+        {7, 13}, {21, 13}, {35, 13}, {49, 13}, {63, 13}, {77, 13}, {91, 13}, {105, 13}, {119, 13}, {133, 13}, {147, 13}, {161, 13}, {175, 13}, {196, 13}, {217, 13},
         {10, 24}, {28, 24}, {42, 24}, {56, 24}, {70, 24}, {84, 24}, {98, 24}, {112, 24}, {126, 24}, {140, 24}, {154, 24}, {168, 24}, {182, 24}, {200, 24}, {217, 24},
-        {12, 35}, {32, 35}, {46, 35}, {60, 35}, {74, 35}, {88, 35}, {102, 35}, {116, 35}, {130, 35}, {144, 35}, {158, 35}, {172, 35}, {194, 35}, {217, 35}, 
-        {16, 45},    {38, 45}, {52, 45}, {66, 45}, {80, 45}, {94, 45}, {108, 45}, {122, 45}, {136, 45}, {150, 45}, {164, 45}, {184, 45}, {203, 45}, {217, 45}, 
-        {9, 56}, {26, 56}, {44, 56}, {96, 56}, {147, 56}, {161, 56}, {175, 56}, {189, 56}, {203, 56}, {217, 56}, 
+        {12, 35}, {32, 35}, {46, 35}, {60, 35}, {74, 35}, {88, 35}, {102, 35}, {116, 35}, {130, 35}, {144, 35}, {158, 35}, {172, 35}, {194, 35}, {217, 35},
+        {16, 45},    {38, 45}, {52, 45}, {66, 45}, {80, 45}, {94, 45}, {108, 45}, {122, 45}, {136, 45}, {150, 45}, {164, 45}, {184, 45}, {203, 45}, {217, 45},
+        {9, 56}, {26, 56}, {44, 56}, {96, 56}, {147, 56}, {161, 56}, {175, 56}, {189, 56}, {203, 56}, {217, 56},
         {30, 5}, {60, 5}, {90, 5}, {120, 5}, {150, 5}, {180, 5}, {210, 5},{210, 32},{210, 60},{180, 60},{150, 60},{120, 60}, {90, 60},{60, 60},{30, 60},{30, 32},
     },
      {
-        4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 
+        4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4,
         4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4,
         4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4,
         4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4,
-        4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 
-        4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 
+        4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4,
+        4, 4, 4, 4, 4, 4, 4, 4, 4, 4,
         2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2,
     }
 };
@@ -112,12 +112,12 @@ void rgb_matrix_indicators_kb(void) {
     if (host_keyboard_led_state().caps_lock) {  // Capslock = RED
         //rgb_matrix_set_color(44, 200, 0, 0);
         rgb_matrix_set_color(46, 200, 0, 0);
-        
-    } 
+
+    }
     //if (host_keyboard_led_state().scroll_lock) {  // Capslock = RED
       //  rgb_matrix_set_color(15, 200, 0, 0);
-        
-   // } 
+
+   // }
 }
 
 #endif  // #ifdef RGB_MATRIX_ENABLE
